[FIX] removed signwarnings
git-svn-id: http://svn.seqan.de/seqan/trunk@12613 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/extras/apps/razers3/razers.h
@@ -555,7 +555,7 @@ enum {
 //                    std::cout << "begin: "<<m.beginPos <<"\tendPos: "<<m.endPos << "\terrors: "<<m.score <<std::endl;
 					appendValue(*matches, m, Generous());
 
-					if (length(*matches) > options->compactThresh)
+					if ((__int64)length(*matches) > options->compactThresh)
 					{
                         double beginTime = sysTime();
 						typename Size<TMatches>::Type oldSize = length(*matches);
